homelab: traffic forwarding, round 3 (with extra ssl)
Attempting to get SSL connections working on the previous setup was...
difficult. The PROXY protocol server doing proxy_pass for SSL connections
couldn't accept SSL traffic, and so tried to read HTTPS traffic as HTTP. This
doesn't work for obvious reasons. So, the solution was to flip the dynamic on
it's head - instead of the proxy_passing endpoints being PROXY protocol-enabled,
make the real endpoints accept PROXY protocol, and have the gateway send traffic
there. For local network requests, create two stream servers - one for HTTP, one
for HTTPS - that just wrap incoming traffic in PROXY protocol and proxy_pass it
back to the real endpoints.

Oh, and we can use Nginx on the gateway side of things as well, so let's keep
things consistent. Now we have working HTTP and HTTPS service routing!

non-http(s) traffic coming soon, again